+++ src/imap_stream.rs
@@ -4,11 +4,11 @@ use async_std::io::{self, Read, Write};
 use async_std::prelude::*;
 use async_std::stream::Stream;
 use async_std::sync::Arc;
+use byte_pool::{Block, BytePool};
 use futures::task::{Context, Poll};
-use imap_proto::{RequestId, Response};
 use nom::Needed;
 
-use byte_pool::{Block, BytePool};
+use crate::types::{Request, ResponseData};
 
 const INITIAL_CAPACITY: usize = 1024 * 4;
 
@@ -17,7 +17,7 @@ lazy_static::lazy_static! {
 }
 
 #[derive(Debug)]
-pub struct ResponseStream<R: Read> {
+pub struct ImapStream<R: Read> {
     // TODO: write some buffering logic
     pub(crate) inner: R,
     buffer: Block<'static>,
@@ -31,10 +31,10 @@ enum DecodeResult {
     None(Block<'static>),
 }
 
-impl<R: Read + Write + Unpin> ResponseStream<R> {
-    /// Creates a new `ResponseStream` based on the given `Read`er.
+impl<R: Read + Write + Unpin> ImapStream<R> {
+    /// Creates a new `ImapStream` based on the given `Read`er.
     pub fn new(inner: R) -> Self {
-        ResponseStream {
+        ImapStream {
             inner,
             buffer: POOL.alloc(INITIAL_CAPACITY),
             current: 0,
@@ -58,7 +58,7 @@ impl<R: Read + Write + Unpin> ResponseStream<R> {
     }
 }
 
-impl<R: Read + Unpin> ResponseStream<R> {
+impl<R: Read + Unpin> ImapStream<R> {
     fn decode(&mut self, buf: Block<'static>, n: usize) -> io::Result<DecodeResult> {
         if self.decode_needs > n {
             return Ok(DecodeResult::None(buf));
@@ -93,7 +93,7 @@ impl<R: Read + Unpin> ResponseStream<R> {
     }
 }
 
-impl<R: Read + Unpin> Stream for ResponseStream<R> {
+impl<R: Read + Unpin> Stream for ImapStream<R> {
     type Item = io::Result<ResponseData>;
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
@@ -143,68 +143,5 @@ impl<R: Read + Unpin> Stream for ResponseStream<R> {
                 }
             }
         }
-    }
-}
-
-rental! {
-    pub mod rents {
-        use super::*;
-
-        #[rental(debug, covariant)]
-        pub struct ResponseData {
-            raw: Block<'static>,
-            response: Response<'raw>,
-        }
-    }
-}
-
-pub use rents::ResponseData;
-
-impl std::cmp::PartialEq for ResponseData {
-    fn eq(&self, other: &Self) -> bool {
-        self.parsed() == other.parsed()
-    }
-}
-
-impl std::cmp::Eq for ResponseData {}
-
-impl ResponseData {
-    pub fn request_id(&self) -> Option<&RequestId> {
-        match self.suffix() {
-            Response::Done { ref tag, .. } => Some(tag),
-            _ => None,
-        }
-    }
-
-    pub fn parsed(&self) -> &Response<'_> {
-        self.suffix()
-    }
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub struct Request(pub Option<RequestId>, pub Vec<u8>);
-
-#[derive(Debug)]
-pub struct IdGenerator {
-    next: u64,
-}
-
-impl IdGenerator {
-    pub fn new() -> Self {
-        Self { next: 0 }
-    }
-}
-
-impl Default for IdGenerator {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Iterator for IdGenerator {
-    type Item = RequestId;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next += 1;
-        Some(RequestId(format!("A{:04}", self.next % 10_000)))
     }
 }
