+++ src/types/id_generator.rs
@@ -1,0 +1,26 @@
+use imap_proto::RequestId;
+
+#[derive(Debug)]
+pub struct IdGenerator {
+    next: u64,
+}
+
+impl IdGenerator {
+    pub fn new() -> Self {
+        Self { next: 0 }
+    }
+}
+
+impl Default for IdGenerator {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Iterator for IdGenerator {
+    type Item = RequestId;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.next += 1;
+        Some(RequestId(format!("A{:04}", self.next % 10_000)))
+    }
+}
